# services/backend/src/main/resources/swagger/greenit/new-workspace.yml
openapi: "3.0.3"
info:
  title: Swagger GreenIT - Open API 3.0 - workspace
  description: This is a Inventory Server based on the OpenAPI 3.0 specification.
  version: "1.0.0"

tags:
  - name: new-workspace
    description: All endpoint for a workspace creation from top header

paths:
  /workspace/workspaces:
    post:
      tags:
        - new-workspace
      summary: Create workspace
      operationId: createWorkspace
      requestBody:
        description: Mandatory workspace details to create.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpsertRest"
      responses:
        "201":
          description: workspace has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceRest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          description: workspace already exists.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/RestError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /workspace/domain-organizations:
    post:
      tags:
        - new-workspace
      summary: get organizations for domain
      operationId: getDomainOrganization
      requestBody:
        description: user's email.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDetailsRest"
      responses:
        "200":
          description: Return organizations information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizationDetailsRest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"